Back-end/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.43
langchain-openai==0.3.8
python-dotenv==1.0.1